+++ NEWS.md
@@ -1,3 +1,25 @@
+# waldo 0.5.0
+
+* You can opt-out of quoting strings with `quote_strings = FALSE` (#145).
+
+* Improvements to missing value handling:
+
+  * `NA_character_` and `"NA"` are no longer treated as equal (#162).
+
+  * `NA_real_` and `NaN` are no longer treated as equal (@sorhawell, #150).
+
+  * Leading and trailing `NA`s are no longer omitted from output when the
+    lengths of `x` and `y` are unequal (#109).
+
+* The `balanced` attribute used by some `POSIXlt` objects in R 4.3 and greater
+  is now ignored (#160).
+
+* 3d (and greater) numeric arrays no longer cause an error (#148).
+
+* Support for complex numbers is improved (#146).
+
+* `ignore_attr = "class"` now works for more types of input (#143).
+
 # waldo 0.4.0
 
 * Atomic S3 classes with format methods now use those methods when